Fix episode list using wrong size for thumbnails
Previous commit changed the size of episode thumbnails to improve their
quality on the home page, but that also made them larger on the episode
list. This makes each part of the app use the appropriate size.

(the episode list could use larger thumbnails eventually, but I'd want
to redesign it to better accommodate them)